% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d}
\alias{d}
\title{General information to peat samples analyzed in the manuscript.}
\format{
A data frame with 59 rows and 8 variables:
\describe{
  \item{id_90}{An integer representing the sample IDs used in the
  manuscript's computations.}
  \item{site_label}{A character vector with the site labels used in the
  manuscript.}
  \item{site_name}{A character vector with the site names.}
  \item{site_order}{A numeric vector representing the order in which sites
  are typically displayed in the manuscript.}
  \item{depth_upper}{A numeric vector with the upper depth of each
  peat layer from the ground surface [cm].}
  \item{depth_lower}{A numeric vector with the lower depth of each
  peat layer from the ground surface [cm].}
  \item{id_ftir}{A character vector with the labels of the mid infrared
  spectra for each peat layer.}
  \item{id_integrate2}{An integer representing the sample IDs used in the
  manuscript's computations (this is a reference ID, additional to id_90).}
}
}
\usage{
d
}
\description{
A dataset containing general information to the peat samples analyzed in
this manuscript (sample IDs, assignments to sites, depth layers, file names
of MIRS files).
}
\keyword{datasets}
