% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_irms}
\alias{d_irms}
\title{Compilation of element content and isotope signature data for the peat samples analyzed in the manuscript.}
\format{
A data frame with 59 rows and 8 variables:
\describe{
  \item{id_90}{An integer representing the sample IDs used in the
  manuscript's computations.}
  \item{C}{The sample C content [g/g].}
  \item{N}{The sample N content [g/g].}
  \item{H}{The sample H content [g/g].}
  \item{O}{The sample O content [g/g].}
  \item{S}{The sample S content [g/g].}
  \item{d15N}{The isotope signature of \eqn{^{15}}N
  (\eqn{\delta^{15}}N\textsubscript{air}) [\textperthousand].}
  \item{d13C}{The isotope signature of \eqn{^{13}}C
  (\eqn{\delta^{13}}C\textsubscript{VPDB}) [\textperthousand].}
}
}
\usage{
d_irms
}
\description{
A dataset containing the compiled element content and isotope signature data for
peat samples used in this study.
}
\keyword{datasets}
