% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_mir}
\alias{d_mir}
\title{Compilation of mid infrared spectra of peat samples analyzed in the manuscript.}
\format{
An object of class \code{\link[ir:ir_new_ir]{ir}} with 59 rows and 4
variables:
\describe{
  \item{id_90}{An integer representing the sample IDs used in the
  manuscript's computations.}
  \item{sample_id}{The same as id_90.}
  \item{measurement_id}{An integer representing an increasing IDs for
  individual measurements.}
  \item{spectra}{The mid infrared spectra}.
}
}
\usage{
d_mir
}
\description{
A dataset containing the compiled raw (unprocessed) and preprocessed mid
infrared spectra for peat samples used in this study. Spectra that could not
be obtained as raw data were included as already baseline corrected spectra.
}
\keyword{datasets}
