% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_sites}
\alias{d_sites}
\title{Information on the sites where peat samples were collected.}
\format{
A data frame with 15 rows and 10 variables:
\describe{
  \item{site_id}{An integer representing a unique ID for each site.}
  \item{site_label}{A character vector with the site labels used in the
  manuscript.}
  \item{site_name}{A character vector with the site names.}
  \item{longitude}{A numeric vector with the longitude for each site.}
  \item{latitude}{A numeric vector with the latitude for each site.}
  \item{altitude}{A numeric vector with the altitude for each site, as
  extracted from the median values of the GMTED2010 raster files
  \insertCite{Danielson.2011}{redoxpeat}.}
  \item{peatland_type}{A character vector with a classification of the sites
  as bogs or fens following available studies for the respective sites or
  from own investigations following concepts in
  \insertCite{Rydin.2013}{redoxpeat}.}
  \item{current_vegetation}{A character vector with species names/plant
  functional groups of the dominant current vegetation at the sites.}
  \item{tmp_mean}{A numeric vector with the mean annual temperature
  [\eqn{^\circ}C] 1970 to 2000 derived from WorldClim version 2.1 climate data
  (30 seconds spatial resolution, monthly temporal resolution)
  \insertCite{Fick.2017}{redoxpeat}.}
  \item{prc_sum}{A numeric vector with the annual total precipitation
  [mm] averaged over 1970 to 2000 derived from WorldClim version 2.1 climate
  data (30 seconds spatial resolution, monthly temporal resolution)
  \insertCite{Fick.2017}{redoxpeat}.}
}
}
\source{
\insertCite{Fick.2017}{redoxpeat}
     \insertCite{Danielson.2011}{redoxpeat}
}
\usage{
d_sites
}
\description{
A dataset containing basic information on the peatland sites where peat
samples were collected from for this study.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
