% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_xrf}
\alias{d_xrf}
\title{Compilation of element content data from wavelength-dispersive X-ray fluorescence spectroscopy for the peat samples analyzed in the manuscript.}
\format{
A data frame with 59 rows and 22 variables:
\describe{
  \item{id_90}{An integer representing the sample IDs used in the
  manuscript's computations.}
  \item{Na}{The sample Na content [µg/g].}
  \item{Mg}{The sample Mg content [µg/g].}
  \item{Al}{The sample Al content [µg/g].}
  \item{Si}{The sample Si content [µg/g].}
  \item{P}{The sample P content [µg/g].}
  \item{S}{The sample S content [µg/g].}
  \item{Cl}{The sample Cl content [µg/g].}
  \item{K}{The sample K content [g/g].}
  \item{Ca}{The sample Ca content [g/g].}
  \item{Ti}{The sample Ti content [µg/g].}
  \item{Cr}{The sample Cr content [µg/g].}
  \item{Mn}{The sample Mn content [µg/g].}
  \item{Fe}{The sample Fe content [g/g].}
  \item{Cu}{The sample Cu content [µg/g].}
  \item{Zn}{The sample Zn content [µg/g].}
  \item{As}{The sample As content [µg/g].}
  \item{Br}{The sample Br content [µg/g].}
  \item{Rb}{The sample Rb content [µg/g].}
  \item{Sr}{The sample Sr content [µg/g].}
  \item{Ba}{The sample Ba content [µg/g].}
  \item{Pb}{The sample Pb content [µg/g].}
}
}
\usage{
d_xrf
}
\description{
A dataset containing the compiled element content data from
wavelength-dispersive X-ray fluorescence spectroscopy for the peat samples
analyzed in the manuscript.
}
\keyword{datasets}
