% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{el_t0}
\alias{el_t0}
\title{EAC and EDC values for the peat samples analyzed in the manuscript.}
\format{
A data frame with 472 rows and 11 variables:
\describe{
  \item{id_90}{An integer representing the sample IDs used in the
  manuscript's computations.}
  \item{site_label}{A character vector with site labels for sites
  where the peat samples were taken.}
  \item{site_label}{A character vector with site names for sites
  where the peat samples were taken.}
  \item{depth_upper}{A numeric vector with the depth of the upper boundary
  of the peat layers from which the samples were taken [cm].}
  \item{measurement_id}{An integer representing the ID of replicate
  measurements, nested within id_90.}
  \item{mass}{The mass of the sample in the suspension [g].}
  \item{C}{The sample C content [g/g].}
  \item{eac}{The measured electron accepting capacity (EAC) [µmol].}
  \item{eac_c}{The EAC normalized to the sample's C content
  [µmol C\eqn{_\text{C}^{-1}}].}
  \item{edc}{The measured electron donating capacity (EDC) [µmol].}
  \item{edc_c}{The EDC normalized to the sample's C content
  [µmol C\eqn{_\text{C}^{-1}}].}
}
}
\usage{
el_t0
}
\description{
A dataset containing the EAC and EDC data for peat samples used in this
study.
}
\keyword{datasets}
