% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fe_t0}
\alias{fe_t0}
\title{Contents of different iron ions and total iron in the peat samples analyzed in the manuscript.}
\format{
A data frame with 59 rows and 6 variables:
\describe{
  \item{id_90}{An integer representing the sample IDs used in the
  manuscript's computations.}
  \item{mass}{The mass of the sample in the suspension [g].}
  \item{fe2}{The concentration of Fe\eqn{^{2+}} in the samples [µmol/g].}
  \item{fe3}{The concentration of Fe\eqn{^{3+}} in the samples [µmol/g].}
  \item{fe_tot}{fe2 + fe3 [µmol/g].}
}
}
\usage{
fe_t0
}
\description{
A dataset containing the concentrations of different iron ions and total iron
as obtained by acid extraction in the analyzed peat samples.
}
\keyword{datasets}
