% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_get_ci.R
\name{get_ci}
\alias{get_ci}
\alias{get_lower_ci}
\alias{get_upper_ci}
\title{Functions to compute lower and upper confidence intervals of a sample.}
\usage{
get_lower_ci(x, prob)

get_upper_ci(x, prob)
}
\arguments{
\item{x}{A numeric vector representing a sample from a distribution for which
to compute a confidence interval limit.}

\item{prob}{A numeric value in [0;1] representing the confidence level of the
confidence intervals to compute.}
}
\value{
The lower (\code{get_lower_ci}) or upper (\code{get_upper_ci}) limit of the
confidence interval (two-sided) for \code{x}.
}
\description{
\code{get_lower_ci} and \code{get_upper_ci} compute the lower and upper confidence
limits (two-sided) for samples from a distribution.
}
\examples{
# get a sample from a distribution
x <- rnorm(100, 5, 2)

# get the lower 95\% confidence limit
x_lwr <- get_lower_ci(x, prob = 0.95)
x_upr <- get_upper_ci(x, prob = 0.95)

}
