% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redoxpeat-package.R
\docType{package}
\name{redoxpeat-package}
\alias{redoxpeat}
\alias{redoxpeat-package}
\title{redoxpeat: Research Compendium for the Manuscript "Electrochemical Properties of 
    Peat Particulate Organic Matter on a Global Scale: Relation to Peat 
    Chemistry and Degree of Decomposition".}
\description{
This package is the research compendium for the manuscript
    "Electrochemical Properties of Peat Particulate Organic Matter on a 
    Global Scale: Relation to Peat Chemistry and Degree of Decomposition".
    It contains all code and data to reproduce the computations made to
    derive results and create figures shown in the manuscript.
}
\author{
\strong{Maintainer}: Henning Teickner \email{h_teic01@uni-muenster.de}

}
\keyword{internal}
