% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tan2017}
\alias{tan2017}
\title{EAC and EDC values for various HS derived from \insertCite{Tan.2017;textual}{redoxpeat}.}
\format{
A data frame with 27 rows and 4 variables:
\describe{
  \item{eac}{A numeric vector with the EAC [mmol/g].}
  \item{edc}{A numeric vector with the EAC [mmol/g].}
  \item{tmp_eac}{A numeric vector with x axis values (average annual
  temperature) in figure 2 for the EAC in
  \insertCite{Tan.2017;textual}{redoxpeat}.}
  \item{tmp_edc}{A numeric vector with x axis values (average annual
  temperature) in figure 2 for the EDC in
  \insertCite{Tan.2017;textual}{redoxpeat}.}
}
}
\source{
\insertCite{Tan.2017;textual}{redoxpeat}
}
\usage{
tan2017
}
\description{
A dataset containing EAC and EDC values for several mineral soil HS as
derived from \insertCite{Tan.2017;textual}{redoxpeat}. Data were
derived from figure 2 in the publication using the R package
\insertCite{Poisot.2011}{redoxpeat}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
