% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{van_krevelen_om}
\alias{van_krevelen_om}
\title{Organic matter fraction regions in van-Krevelen diagrams derived from \insertCite{Kim.2003;textual}{redoxpeat}.}
\format{
A data frame with 5 rows and 6 variables:
\describe{
  \item{om_fraction}{A character with the name of the organic matter
  fraction.}
  \item{angle}{A numeric vector with the angle of the ellipse denoting the
  area of the OM fraction.}
  \item{O:C}{A numeric vector with the position of the ellipse denoting the
  area of the OM fraction along the O:C axis.}
  \item{H:C}{A numeric vector with the position of the ellipse denoting the
  area of the OM fraction along the H:C axis.}
  \item{oc_halfwidth}{A numeric vector with the half width of the ellipse
  denoting the area of the OM fraction along the O:C axis.}
  \item{hc_halfwidth}{A numeric vector with the half width of the ellipse
  denoting the area of the OM fraction along the H:C axis.}
}
}
\source{
\insertCite{Kim.2003;textual}{redoxpeat}
}
\usage{
van_krevelen_om
}
\description{
A dataset containing data to create ellipses representing regions
characteristic for different organic matter fractions in van-Krevelen
diagrams as derived from \insertCite{Kim.2003;textual}{redoxpeat}. Data were
derived from figure 5 in the publication by estimating positions and angles
of ellipses by eye.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
