% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{walpen2018}
\alias{walpen2018}
\title{EAC and EDC values for various DOM samples derived from \insertCite{Walpen.2018;textual}{redoxpeat}.}
\format{
A data frame with 111 rows and 6 variables:
\describe{
  \item{eac_mean}{A numeric vector with the EAC [mmol/g\eqn{_\text{C}}].}
  \item{edc_mean}{A numeric vector with the EAC [mmol/g\eqn{_\text{C}}].}
  \item{suva}{A numeric vector with x axis values
  (SUVA\textsubscript{254} [L mg\eqn{_\text{C}^{-1}} m\eqn{^{-1}}]) in
  figure 3 for the EAC and EDC in
  \insertCite{Walpen.2018;textual}{redoxpeat}.}
  \item{y}{A numeric vector with y axis values (EAC and EDC
  [mmol/g\eqn{_\text{C}}]) in figure 3 for
  the EDC in \insertCite{Walpen.2018;textual}{redoxpeat}.}
  \item{sample_type}{A character vector representing the sample type
  (pore water or pool water).}
  \item{comment}{A character vector with a comment how the EAC/EDC values
  were normalized.}
}
}
\source{
\insertCite{Walpen.2018;textual}{redoxpeat}
}
\usage{
walpen2018
}
\description{
A dataset containing EAC and EDC values for several peat DOM samples as
derived from \insertCite{Walpen.2018;textual}{redoxpeat}. Data were
derived from figure 3 in the publication using the R package
\insertCite{Poisot.2011}{redoxpeat}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
