% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redoxpeat-package.R
\docType{package}
\name{redoxpeat-package}
\alias{redoxpeat}
\alias{redoxpeat-package}
\title{redoxpeat: Reproducible research compendium with R code and data for: 
    'Electrochemical Properties of Peat Particulate Organic Matter on a Global 
    Scale: Relation to Peat Chemistry and Degree of Decomposition'}
\description{
This package is the research compendium for the manuscript
    "Electrochemical Properties of Peat Particulate Organic Matter on a 
    Global Scale: Relation to Peat Chemistry and Degree of Decomposition".
    It contains all code and data to reproduce the computations made to
    derive results and create figures shown in the manuscript.
}
\author{
\strong{Maintainer}: Henning Teickner \email{henning.teickner@uni-muenster.de} (\href{https://orcid.org/0000-0002-3993-1182}{ORCID})

Authors:
\itemize{
  \item Chuanyu Gao \email{gaochuanyu@iga.ac.cn}
  \item Klaus-Holger Knorr \email{kh.knorrr@wwu.de}
}

}
\keyword{internal}
